refactor: adjust pkg directory of generator and patcher
