+++ pkg/generator/appconfiguration/generator/app_configurations_generator.go
@@ -6,8 +6,11 @@ import (
 	"kusionstack.io/kusion/pkg/generator"
 	"kusionstack.io/kusion/pkg/generator/appconfiguration"
 	accessories "kusionstack.io/kusion/pkg/generator/appconfiguration/generator/accessories/database"
+	"kusionstack.io/kusion/pkg/generator/appconfiguration/generator/monitoring"
 	"kusionstack.io/kusion/pkg/generator/appconfiguration/generator/trait"
 	"kusionstack.io/kusion/pkg/generator/appconfiguration/generator/workload"
+	patmonitoring "kusionstack.io/kusion/pkg/generator/appconfiguration/patcher/monitoring"
+	pattrait "kusionstack.io/kusion/pkg/generator/appconfiguration/patcher/trait"
 	"kusionstack.io/kusion/pkg/models"
 	appmodel "kusionstack.io/kusion/pkg/models/appconfiguration"
 	"kusionstack.io/kusion/pkg/projectstack"
@@ -96,7 +99,7 @@ func (g *appConfigurationGenerator) Generate(spec *models.Spec) error {
 		accessories.NewDatabaseGeneratorFunc(g.project, g.stack, g.appName, g.app.Workload, g.app.Database),
 		workload.NewWorkloadGeneratorFunc(g.project, g.stack, g.appName, g.app.Workload),
 		trait.NewOpsRuleGeneratorFunc(g.project, g.stack, g.appName, g.app),
-		NewMonitoringGeneratorFunc(g.project, g.app.Monitoring, g.appName),
+		monitoring.NewMonitoringGeneratorFunc(g.project, g.app.Monitoring, g.appName),
 		// The OrderedResourcesGenerator should be executed after all resources are generated.
 		NewOrderedResourcesGeneratorFunc(),
 	}
@@ -106,8 +109,8 @@ func (g *appConfigurationGenerator) Generate(spec *models.Spec) error {
 
 	// Patcher logic patches generated resources
 	pfs := []appconfiguration.NewPatcherFunc{
-		trait.NewOpsRulePatcherFunc(g.app),
-		NewMonitoringPatcherFunc(g.appName, g.app, g.project),
+		pattrait.NewOpsRulePatcherFunc(g.app),
+		patmonitoring.NewMonitoringPatcherFunc(g.appName, g.app, g.project),
 	}
 	if err := appconfiguration.CallPatchers(spec.Resources.GVKIndex(), pfs...); err != nil {
 		return err
