+++ pkg/generator/appconfiguration/generator/monitoring/monitoring_generator.go
@@ -1,4 +1,4 @@
-package generator
+package monitoring
 
 import (
 	"fmt"
