+++ pkg/generator/appconfiguration/generator/monitoring/monitoring_generator_test.go
@@ -1,4 +1,4 @@
-package generator
+package monitoring
 
 import (
 	"fmt"
