+++ pkg/generator/appconfiguration/patcher/monitoring/monitoring_patcher.go
@@ -1,4 +1,4 @@
-package generator
+package monitoring
 
 import (
 	appsv1 "k8s.io/api/apps/v1"
