+++ pkg/generator/appconfiguration/patcher/monitoring/monitoring_patcher_test.go
@@ -1,4 +1,4 @@
-package generator
+package monitoring
 
 import (
 	"fmt"
