+++ pkg/generator/appconfiguration/patcher/trait/ops_rule_patcher.go
