+++ pkg/generator/appconfiguration/patcher/trait/ops_rule_patcher_test.go
